Preserve mixed optional memo deps
